Move home page sections to separate components
